Moved saved posts from sidebar to profile
