+++ src/pages/Profile/index.css
@@ -16,6 +16,46 @@
   margin: auto;
 }
 
+.feed_btn_container {
+  margin-left: 17vw;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 40px;
+  padding: 10px;
+}
+
+.feed_change_btn {
+  font-size: 20px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding-top: 15px;
+    padding-bottom: 10px;
+    padding-inline: 20px;
+    cursor: pointer;
+    background: var(--bg-color);
+    border: none;
+    color: var(--text-secondary);
+}
+
+.feed_btn_deactivated {
+  opacity: 0.7;
+  color: var(--profile-color);
+}
+
+.feed_btn_deactivated:hover {
+  opacity: 1;
+}
+
+.feed_btn_activated {
+  border-top: 4px solid;
+}
+
+.feed_btn_text {
+  margin-left: 8px;
+}
+
 .outer-profile-box {
   position: relative;
   margin-top: 0;
@@ -278,10 +318,26 @@
   }
 }
 
+@media screen and (max-width: 1200px) {
+  .background-image-container {
+    padding: 4em 0 0 8vw;
+  }
+
+  .feed_btn_container {
+    margin-left: 8vw;
+  }
+}
+
+
 @media screen and (max-width: 800px) {
   .outer-profile-box {
     width: calc(100% - 10vw);
   }
+
+  .feed_btn_container {
+    margin-left: 0;
+  }
+
   .background-image-container {
     padding: 0 !important;
     padding-top: 4em !important;
@@ -304,12 +360,6 @@
   .bg-save-btn {
     top: 60px;
     margin-top: 0;
-  }
-}
-
-@media screen and (max-width: 1200px) {
-  .background-image-container {
-    padding: 4em 0 0 8vw;
   }
 }
