+++ src/pages/Profile/index.jsx
@@ -13,6 +13,9 @@ import {
 import { lazy, useEffect, useState } from "react";
 import { playErrorSound, playSuccessSound } from "../../js/sounds";
 import { useNavigate, useParams } from "react-router-dom";
+import GridOnIcon from '@mui/icons-material/GridOn';
+import BookmarksIcon from "@mui/icons-material/Bookmarks";
+import { doc, getDoc } from "firebase/firestore";
 
 import Cam from "@mui/icons-material/CameraAltOutlined";
 import EditIcon from "@mui/icons-material/Edit";
@@ -41,6 +44,7 @@ function Profile() {
 
   const [user, setUser] = useState(null);
   const [feed, setFeed] = useState([]);
+  const [savedPosts, setSavedPosts] = useState([]);
   const [open, setOpen] = useState(false);
   const [friendRequestSent, setFriendRequestSent] = useState(false);
   const [userData, setUserData] = useState(null);
@@ -50,6 +54,7 @@ function Profile() {
   const [editing, setEditing] = useState(false);
   const [bgimgurl, setBgimgurl] = useState(null);
   const [backgroundImage, setBackgroundImage] = useState(null);
+  const [showSaved, setShowSaved] = useState(false);
 
   let name = "";
   let avatar = "";
@@ -87,7 +92,7 @@ function Profile() {
           .put(backgroundImage);
         uploadTask.on(
           "state_changed",
-          () => {},
+          () => { },
           (error) => {
             enqueueSnackbar(error.message, {
               variant: "error",
@@ -319,6 +324,38 @@ function Profile() {
       unsubscribe();
     };
   }, [user, name]);
+
+
+  async function getSavedPosts() {
+    let savedPostsArr = JSON.parse(localStorage.getItem("posts")) || [];
+    const posts = [];
+    if (savedPostsArr.length > 0 && savedPosts.length === 0) {
+      try {
+        const fetchSavedPosts = savedPostsArr.map(async (id) => {
+          try {
+            const docRef = doc(db, "posts", id);
+            const docSnap = await getDoc(docRef);
+            if (docSnap.exists()) {
+              posts.push({ id: docSnap.id, post: docSnap.data() });
+            }
+          } catch (error) {
+            console.log(error, "err")
+            enqueueSnackbar("Error while fetching all posts", {
+              variant: "error",
+            });
+          }
+        });
+
+        await Promise.all(fetchSavedPosts); // Wait for all fetch requests to complete
+        setSavedPosts(posts);
+      } catch (error) {
+        console.log(error, "err2")
+        enqueueSnackbar("Error while fetching all posts", {
+          variant: "error",
+        });
+      }
+    }
+  }
 
   return (
     <>
@@ -557,7 +594,26 @@ function Profile() {
               </Box>
             </Box>
           </Box>
-          <ProfieFeed feed={feed} user={user} />
+          <div className="feed_btn_container">
+            <button
+              className={`feed_change_btn ${showSaved ? "feed_btn_deactivated" : "feed_btn_activated"}`}
+              onClick={() => setShowSaved(false)}
+            >
+              <GridOnIcon /> <span className="feed_btn_text">Feed</span>
+            </button>
+            {user?.uid === uid && (
+              <button
+                className={`feed_change_btn ${showSaved ? "feed_btn_activated" : "feed_btn_deactivated"}`}
+                onClick={() => {
+                  getSavedPosts()
+                  setShowSaved(true)
+                }}
+              >
+                <BookmarksIcon /> <span className="feed_btn_text">Saved</span>
+              </button>
+            )}
+          </div>
+          {showSaved? (<ProfieFeed feed={savedPosts}/>) : (<ProfieFeed feed={feed} />)}
         </>
       ) : userExists ? (
         <Loader />
